[feat] add comments for each kernel and dump kernel with meta

+++ cpp/include/merlin/util.cuh
@@ -49,14 +49,12 @@ __inline__ __device__ int64_t atomicExch(int64_t* address, int64_t val) {
 }
 
 __inline__ __device__ int64_t atomicAdd(int64_t* address, const int64_t val) {
-  return (int64_t)atomicAdd((unsigned long long*)address,
-                            (const unsigned long long)val);
+  return (int64_t)atomicAdd((unsigned long long*)address, val);
 }
 
 __inline__ __device__ uint64_t atomicAdd(uint64_t* address,
                                          const uint64_t val) {
-  return (uint64_t)atomicAdd((unsigned long long*)address,
-                             (const unsigned long long)val);
+  return (uint64_t)atomicAdd((unsigned long long*)address, val);
 }
 
 namespace nv {
